packages/authkit/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "fast_authkit"
version = "0.1.0"
description = "Django-like authentication toolkit for FastAPI with JWT."
readme = "../../Readme.md"
requires-python = ">=3.10"
license = { file = "../../LICENSE" }
authors = [{ name = "Azim Hossain Tuhin" }]
dependencies = [
  "fastapi>=0.110",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.urls]
Repository = "https://example.com"

[tool.hatch.build.targets.wheel]
packages = ["src/authkit"]
